publisher/serverservice: update attribute, versioned attribute NS
Each component data is now stored with a namespace value
which includes either out-of-band, or inband based on the collection mechanism.

And based on the namespace, the data collected from the device is only compared with
the data from server service that matches the collection mechanism - based on this namespace value.