collect/outofband: split collect() into bmcLogin(), bmcLogout() and bmcInventory
The split is for readability and separation of metrics collected on BMC errors.

Along with this, we ensure that bmcLogout() always executed for each connection
when theres been a context cancellation because of a SIGTERM.
This ensures that when Alloy is given a SIGTERM, given time, it will
terminateall active BMC sessions.