cmd/inband: get inventory asset when --asset-source is specified
Instead of attempting to query the asset-source each time
only query the asset source when the flag is passed in.

This allows alloy to be run on a host without requiring an asset-source
to be available.

The change also prevents assets from being published to serverService
when --asset-source was not set to serverService.