publisher/serverservice: publish vendor, model, serial only when empty in inventory
This is to prevent the inband and out of band collectors
from overwriting what they find as the vendor, model, serial attributes.

For example on the dells, the serial is reported differently,

inband: .J0S2LM3.MXWSJ00214015M.
outofband: J0S2LM3

And so whichever one is reported first will be kept.