publisher/serverservice: rework vendor attributes handling
The device vendor attributes should be added and updated in
serverservice based on a few cases, this reworks the vendor attributes
updating to cover those cases and adds tests for them.

This also fixes a case where Alloy would attempt to Create
vendor attributes instead of Updat'ing existing vendor attributes.