Update rivets to latest release
We've been pinned to an *unreleased* commit of rivets for a long time
now. I blame go for not making this obvious. I decided to fix this now
so that we can back on track and only use merged code paths. Getting the
rivets PR merged and then updating this line comes next.