Update Go to latest version
The action is able to read the go.mod version to discover what go
version to install. So lets use that just like in ironlib to avoid
needing to keep multiple files in sync.

I got rid of GOVERSION used by goreleaser since its unnecessary, go
embeds this info automatically. This is already being used, goreleaser
was just overriding it.

I also switched codeql-analysis to avoid autobuild because it failed
trying to install go1.22 and we already have a way to build the binary,
so lets re use it.