FS-1518: update bmc-toolbox/common (#211)
#### What does this PR do
Update the version of `bmc-toolbox/common` to account for SMC
`x12sth-sys`

#### How can this change be tested by a PR reviewer?
Build alloy at this SHA. Reserve a `m3.small.x86.v1 (w/iGPU)`. Enroll
into the sandbox and follow the inventory. The correct output should be:
```
➜  mctl git:(main) ✗ ./mctl --config ~/.mctl/kind.yaml get server -s abcdf392-1aed-4070-b33c-3394319f4416
[
  {
    "updated": "2024-06-11T19:35:17.898328Z",
    "id": "abcdf392-1aed-4070-b33c-3394319f4416",
    "facility": "sandbox",
    "name": "abcdf392-1aed-4070-b33c-3394319f4416",
    "bmc_address": "10.197.132.37",
    "vendor": "supermicro",
    "model": "x12sth-sys",
    "serial": "E895731X4500534"
  }
]```

Note that the model is tracked as `x12sth-sys` and not the motherboard model of `SYS-510T-MR2-EI018`.