update dependencies
Like in other controllers, here we do a general dependency update as
well as explicitly update to metal-toolbox versions of bmclib,
bmc-common and rivets/v2.

Unlike the flasher case we maintain the current (if suboptimal) usage of
the FleetDB API, because we don't have to tackle that immediately
because of removing the rivets fleetdb module.